# Some further optimizations on inliingbitset _.Map