This gives the following results for me:
```
|            Method |     Mean |    Error |   StdDev |  Gen 0 |  Gen 1 | Allocated |
|------------------ |---------:|---------:|---------:|-------:|-------:|----------:|
|         BitSetMap | 29.76 us | 0.033 us | 0.027 us | 5.8289 | 0.6714 |     48 KB |
| InliningBitSetMap | 23.40 us | 0.256 us | 0.240 us | 3.8147 | 0.3662 |     31 KB |

// * Hints *
Outliers
  Benchmarks.BitSetMap: Default -> 2 outliers were removed (30.29 us, 30.55 us)

```

Overall, I managed to make it inline using new `InlineIfLambda` as well as extracting some code to vars and switching a stack for a System.Collections.Generic.List (aka ResizeArray)